Expand to include manifest
This pushes the :latest tag as a manifest to include multiple
platforms.

We expand to include a ppc64le architecture at the same time.

For safety, docker login/push aren't enabled on pull requests
or non-master branch. Environment variables are currently
protected however this prevents failures on PRs.

The purpose is to include provide a test for go-dockerclient
for docker API 1.34 "Create an Image" tests that has platform
as an argument.